Refactor chain, token, and route config management into build-time script
