+++ src/consts/ibcRoutes.ts
@@ -1,4 +1,4 @@
-import { IbcRoute, IbcToWarpRoute } from '../features/tokens/routes/types';
+import { IbcRoute, IbcToWarpRoute } from '../features/routes/types';
 
 // Configs for manually-defined IBC-only routes
 export const ibcRoutes: Array<IbcRoute | IbcToWarpRoute> = [];
