+++ src/context/README.md
@@ -1,0 +1,4 @@
+### About
+
+This folder contains pre-validated and processed configs for the the chains, tokens, and routes.
+The contents are auto-generated by the `yarn build:configs` command. Changes will be overridden on new builds.
