+++ src/features/transfer/TransferTokenForm.tsx
@@ -18,15 +18,15 @@ import { logger } from '../../utils/logger';
 import { getTokenAddress, isNonFungibleToken } from '../caip/tokens';
 import { ChainSelectField } from '../chains/ChainSelectField';
 import { getChainDisplayName } from '../chains/utils';
+import { useRouteChains } from '../routes/hooks';
+import { RoutesMap, WarpRoute } from '../routes/types';
+import { getTokenRoute, isIbcOnlyRoute } from '../routes/utils';
 import { useStore } from '../store';
 import { SelectOrInputTokenIds } from '../tokens/SelectOrInputTokenIds';
 import { TokenSelectField } from '../tokens/TokenSelectField';
 import { useIsApproveRequired } from '../tokens/approval';
 import { useDestinationBalance, useOriginBalance } from '../tokens/balances';
 import { getToken } from '../tokens/metadata';
-import { useRouteChains } from '../tokens/routes/hooks';
-import { RoutesMap, WarpRoute } from '../tokens/routes/types';
-import { getTokenRoute, isIbcOnlyRoute } from '../tokens/routes/utils';
 import { useAccountAddressForChain, useAccounts } from '../wallet/hooks/multiProtocol';
 
 import { TransferFormValues } from './types';
