+++ src/scripts/buildConfigs/utils.ts
@@ -1,0 +1,6 @@
+import fs from 'fs';
+import { parse } from 'yaml';
+
+export function readYaml(path: string) {
+  return parse(fs.readFileSync(path, 'utf-8'));
+}
